backend/pyproject.toml
[project]
name = "Projects-App"
version = "0.1.0"
description = "Project, task management"
authors = [
  { name = "Jithu Sunny", email = "jithusunnyk@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "asyncpg>=0.29,<1.0",
  "python-dotenv>=1.0,<2.0",
  "alembic>=1.13,<2.0",
  "pydantic>=2.6,<3.0"
]
